+++ src/enum.ts
@@ -1,5 +1,4 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 export enum Commands {
-  SHOW_FILE_DETAILS = 'data-size-count.showFileDetails',
-  SHOW_HIDE_STATUSBAR = 'data-size-count.showHideStatusBar',
+  SHOW_HIDE_STATUSBAR = "data-size-count.showHideStatusBar",
 }