+++ src/types.ts
@@ -1,0 +1,9 @@
+export type RegexCountItem = {
+  order: number;
+  format: string;
+  description?: string;
+  title?: string;
+  regex: string;
+  onlyOnSelection?: boolean;
+  alwaysShow?: boolean;
+};